libhwc1: ifdef out all virtual display code
It doesn't work. There are only errors flooding logs and other associated
issues.

In device trees, the flag, that is used here, for enabling virtual display
was commented out, with a comment saying that virtual display support
requires platform changes. If that's correct at all, virtual display
support can't work, or if it isn't, I suppose it would work if there was
properly working SoC-specific code for it which is unavailable.

Change-Id: I5055ba85d874dc7f09a68762adfaf3875afaef0a